requirements.txt
flask
requests
beautifulsoup4
PyPDF2
python-docx
